Implement L0 protocol version 2
This version adds another field to the header to encode the time (in
nanoseconds since the Unix epoch) of the very first FPGA sample (the one
with fpga_count=0) at offset 8, just before "fpga_count". This increases
the length of the header by 8 bytes and pushes back the offsets of all
the subsequent fields.